+++ chaoscf/__init__.py
@@ -18,7 +18,7 @@ import urllib3
 
 urllib3.disable_warnings()
 
-__version__ = '0.5.0'
+__version__ = '0.5.1'
 __all__ = ["__version__", "auth", "discover"]